pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seeknal"
version = "2.5.0"
description = "All-in-one platform for data and AI/ML engineering"
authors = [
    {name = "Fitra Kacamarga"}
]
readme = "README.md"
requires-python = ">=3.11"
scripts = { seeknal = "seeknal.cli.main:main" }
dependencies = [
    "sqlglot~=25.0.0",
    "croniter>=3.0.0",
    "delta-spark==3.2.0",
    "duckdb>=1.1.3",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "libsql-experimental>=0.0.41",
    "mack>=0.5.0",
    "pandas>=1.3.0",
    "pendulum>=3.0.0",
    "psycopg2-binary>=2.9.0",
    "cryptography>=46.0.5",
    "pyarrow>=18.1.0",
    "pydantic>=2.0.0",
    "pyiceberg>=0.8.1",
    "pyspark>=3.0.0",
    "python-box>=7.3.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "quinn>=0.10.3",
    "s3fs>=2024.12.0",
    "sqlalchemy>=1.4.0",
    "sqlalchemy-libsql>=0.1.0",
    "sqlmodel>=0.0.22",
    "tabulate>=0.9.0",
    "toml>=0.10.0",
    "tenacity>=9.0.0",
    "typer>=0.13.1",
    "jinja2>=3.1.0",
    "pymysql>=1.1.0",
    "pyyaml>=6.0.0",
    "simple-term-menu>=1.6.0",
]

[project.optional-dependencies]
# Atlas Data Platform integration for governance, lineage, and API features
# NOTE: Temporarily disabled - atlas-data-platform is not available in PyPI
# To enable when the package is available:
# atlas = [
#     "atlas-data-platform>=0.1.0",
#     "uvicorn[standard]>=0.27.0",
# ]

# Prefect orchestration for scheduled pipeline execution
prefect = [
    "prefect>=3.1.10,<4.0",
]

# Seeknal Ask — NL-to-Analysis Agent
ask = [
    "deepagents>=0.4.12",
    "langchain>=0.3.0",
    "langgraph>=0.4.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.3.0",
    "rich>=13.0.0",
    "ruamel.yaml>=0.18.0",
    # Data science / ML / analytics libraries for execute_python sandbox
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "seaborn>=0.13.0",
    "statsmodels>=0.14.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "filelock>=3.0",
]

# Gateway — WebSocket server, Telegram bot, SSE streaming
gateway = [
    "seeknal[ask]",
    "starlette>=0.50.0",
    "uvicorn>=0.40.0",
    "python-telegram-bot>=21.0",
    "sse-starlette>=2.0",
    "redis>=5.0",
]

# Temporal worker for seeknal ask
temporal = [
    "seeknal[ask]",
    "temporalio>=1.5.0",
]

# Development dependencies
dev = [
    "black>=24.10.0",
    "icecream>=2.1.3",
    "pytest>=8.3.4",
]

# All optional dependencies
all = [
    "seeknal[dev]",
    "seeknal[prefect]",
    # Note: atlas requires internal package atlas-data-platform, install separately with:
    # uv pip install -e ".[atlas]" when the package is available
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "icecream>=2.1.3",
    "pytest>=8.3.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/seeknal"]

[tool.hatch.build.targets.sdist]
include = ["src/seeknal", "LICENSE", "README.md", "pyproject.toml"]

# Pytest configuration
[tool.pytest.ini_options]
# Test paths and discovery
testpaths = ["tests"]
pythonpath = ["src"]

# Test output options
addopts = [
    "-v",                    # verbose output
    "--tb=short",            # shorter traceback format
    "--strict-markers",       # error on unknown markers
    "--disable-warnings",    # disable warnings during test runs
]

# Ignore slow tests by default
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]
